Bump xplat moving the fetchers
