+++ components/profile/EditProfileModal.tsx
@@ -13,9 +13,9 @@ import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { queryClient } from '../../App';
 import { userAtom } from '../../utils/atoms';
-import { FetchedUser } from '../../utils/queries';
 import { DebounceSession } from '../../utils/utils';
 import { validBio, validDisplayname } from '../../xplat/api';
+import { FetchedUser } from '../../xplat/types';
 import ChangeEmail from './ChangeEmail';
 import ChangePassword from './ChangePassword';
 import UploadImage from './UploadImage';
