+++ components/route/SearchableRouteContainer.tsx
@@ -1,7 +1,7 @@
 import { Box, Divider, ScrollView, Spinner, VStack } from 'native-base';
 import { useEffect, useState } from 'react';
-import { FetchedRoute } from '../../utils/queries';
 import { DebounceSession } from '../../utils/utils';
+import { FetchedRoute } from '../../xplat/types';
 import SearchBar from '../searchbar/SearchBar';
 import RouteRow from './RouteRow';
 
@@ -29,7 +29,11 @@ const SearchableRouteContainer = ({ fetchedRoutes }: Props) => {
       fetchedRoutes.forEach((route, index) => {
         // Here we build a corpus. This is the search text that we should query on
         // For now, we are simply concatenating the name, rating, and all tags
-        const corpus = [route.name, route.grade, route.stringifiedTags]
+        const corpus = [
+          route.name,
+          route.gradeDisplayString,
+          route.stringifiedTags,
+        ]
           .join('$')
           .toLowerCase();
         if (corpus.includes(query)) newDisplayedRouteIndices.push(index);
