+++ screens/profile/Follows.tsx
@@ -10,8 +10,8 @@ import { useState } from 'react';
 import { useQuery } from 'react-query';
 import FollowList from '../../components/profile/FollowList';
 import SearchBar from '../../components/searchbar/SearchBar';
-import { buildUserFetcherFromDocRefId } from '../../utils/queries';
 import { TabGlobalScreenProps } from '../../utils/types';
+import { User } from '../../xplat/types';
 
 type UserTab = 'followers' | 'following';
 
@@ -28,7 +28,7 @@ const Follows = ({ route }: TabGlobalScreenProps<'Follows'>) => {
 
   const { isLoading, isError, data, error } = useQuery(
     userDocRefId,
-    buildUserFetcherFromDocRefId(userDocRefId)
+    User.buildFetcherFromDocRefId(userDocRefId)
   );
 
   if (isLoading) {
