+++ screens/profile/Sends.tsx
@@ -1,18 +1,9 @@
-import { User } from '../../xplat/types/user';
-import { useState } from 'react';
-import { useEffect } from 'react';
-import { Send } from '../../xplat/types/send';
 import { View } from 'native-base';
+import { User } from '../../xplat/types/user';
 
 const Sends = (user: User) => {
-  const [_sends, setSends] = useState<Send[]>();
-
-  useEffect(() => {
-    const getSends = async () => {
-      await user.getSends().then(setSends);
-    };
-    getSends();
-  }, [user]);
+  // TODO
+  console.log('TODO: Implements sends', user.getId());
 
   return <View />;
 };
