+++ screens/route/RouteView.tsx
@@ -22,12 +22,11 @@ import LikeButton from '../../components/misc/LikeButton';
 import UserTag from '../../components/profile/UserTag';
 import RatingModal from '../../components/route/RatingModal';
 import { userAtom } from '../../utils/atoms';
-import { buildRouteFetcherFromDocRefId } from '../../utils/queries';
 import {
   TabGlobalNavigationProp,
   TabGlobalScreenProps,
 } from '../../utils/types';
-import { Post, QueryCursor, RouteStatus } from '../../xplat/types';
+import { Post, QueryCursor, Route, RouteStatus } from '../../xplat/types';
 
 const FORCED_THUMBNAIL_HEIGHT = 200;
 
@@ -54,7 +53,7 @@ const RouteView = ({ route }: TabGlobalScreenProps<'RouteView'>) => {
 
   const { isLoading, isError, data, error } = useQuery(
     routeDocRefId,
-    buildRouteFetcherFromDocRefId(routeDocRefId)
+    Route.buildFetcherFromDocRefId(routeDocRefId)
   );
 
   useEffect(() => {
@@ -118,7 +117,7 @@ const RouteView = ({ route }: TabGlobalScreenProps<'RouteView'>) => {
           <HStack flexWrap="wrap" justifyContent="space-between" mx={4}>
             <Heading size="2xl">{data.name}</Heading>
             <Heading size="2xl" color="grey">
-              {data.grade}
+              {data.gradeDisplayString}
             </Heading>
           </HStack>
           <Text fontSize="2xl" color="grey" mx={4}>
